+++ build.sh
@@ -3,11 +3,22 @@
 rm -rf dist
 mkdir dist
 
-echo "Run dep ensure"
-$GOPATH/bin/dep ensure
-$GOPATH/bin/dep check
+# generate version number
+if [ -n "${TRAVIS_TAG:-}" ]; then
+    VERSION_BASE=${TRAVIS_TAG}  
+else 
+    VERSION_BASE=$(git describe --abbrev=0 --tags 2> /dev/null || echo 'dblp.0.0.0')
+fi
+VERSION_SHORT_COMMIT=$(git rev-parse --short HEAD)
+VERSION_FULL_COMMIT=$(git rev-parse HEAD)
 
-echo "Build tide-whisperer"
-go build -o dist/tide-whisperer tide-whisperer.go
+GO_COMMON_PATH="github.com/tidepool-org/go-common"
+	
+echo "Build tide-whisperer $VERSION_BASE+$VERSION_FULL_COMMIT"
+go mod tidy
+go build -ldflags "-X $GO_COMMON_PATH/clients/version.ReleaseNumber=$VERSION_BASE \
+    -X $GO_COMMON_PATH/clients/version.FullCommit=$VERSION_FULL_COMMIT \
+    -X $GO_COMMON_PATH/clients/version.ShortCommit=$VERSION_SHORT_COMMIT" \
+    -o dist/tide-whisperer tide-whisperer.go
 cp start.sh dist/
 cp env.sh dist/